improve overall coverage report
mostly move tests to separate files
